requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cuser==2017.3.16
django-environ==0.11.2
django-tinymce==4.1.0
modeltranslation==0.25
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
